Start working on two stage running
In order to have a two stage execution we need to have a command we can
run the container with that does nothing and takes no resources (which
will do by mapping the executable itself into the container and using a
hidden command line flag).

So now we have command line flags, and we configure the parser so that
options _have_ to come before arguments, since we want the arguments to
become the command to run in the container. So "wr -i name hello" means
run the command "hello" in a container from the image "name", whereas
"we hello -i name" means run the command "hello -i name" in a container
from the default image.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>