Use the name crate instead of image
What we are defining are basically containers, and we want to be able to
refer to the docker image that the container will be based on. So we now
use the name crate to refer to the wharf rat analogue of a docker
container - which is basically a container with some config.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>