Change the way we name containers
Instead of using a hash of the config to name the container, we just use
the project path and crate name. We then store the config used to create
the container in a label and check it. This way we can detect when the
config has changed since the container was created (previously we used
to just think that there wasn't a container).

The user can use the --clean flag to rebuild the container when this
happens.

Also add some more config and other bits and pieces.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>