Restrucure command code
The main file was getty messy, and there wasn't much structure to the
command code. Now it's got more structure, and we also have the ability
to setup a new wharfrat command - which we will use to provide a set of
management sub-commands. We are also now using sub-commands for the
wr-init command.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>