Add a list sub-command to management mode
Add a list sub command to the wharfrat management command that lists the
wharfrat containers that have been created. For each container it lists
the project path and crate name - coloured red if they don't exist, and
the crate yellow if the config doesn't match. It also lists the
container image and state.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>