Run setup as a separate step
Running the setup as part of the server command meant that we couldn't
actually tell when the setup was finished. Now we run the setup as a
separate step after we have created the continer. This way we can be
sure that the setup has finished before we try running the provided
command.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>