Add vc branch into things more fully
Whilst some people check out branches into different folders, some
don't. We want to have a different container for different branches, so
that we don't keep destroying and creating the containers if someone is
switching between two branches with different configs.

So now, the branch name is included in the container name hash, and
recorded in the container metadata. The list command will look at the
version of the project file for the named branch if the current checkout
points at a different branch (this means that list will not claim that a
container is out of date, when a different branch is checked out).

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>