Check project file in VC exists first
Previously we tried to parse the project file from VC before checking if
the file actually existed - which caused list to exit with an error. Now
we check the existence first, and just mark the project and crate as
red without actually trying to parse the non-existant file (and thus
avoiding the error).

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>