Add command to remove old containers
If a container refers to a crate that cannot be found, then the new
prune command will remove it (provided that it is run with --yes). This
makes it easy to clean up containers from old checkouts/branches (though
if the branch still exists then the container will stay - the tool
doesn't know which branches are considered obsolete).

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>