Add ability to control working dir
Can be either an absolute path (i.e. starts with /), or one of a number
of keywords.

The current keywords are:

  home: The user's home directory
  match: The current directory of the command itself

The default if not set is match.

It is also planned to add (but it hasn't been yet):

  project: Map the current directory into the project

Finally, the plan is to allow a comma separated list, so that if one
entry doesn't work (e.g. home but unable to get a home directory) then
the next will be tried.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>