Add workaround for docker terminal size issue
When using docker exec the terminal size is not set reliably. This is
due to the fact that we can't ensure the terminal size is set before the
command is run as the exec create call also starts the command.

To workaround this issue we reuse the proxy introduced as a workaround
for the workdir support being missing in earlier API versions. So now
the startup sequence when using a terminal is (where a & b are used to
indicate events that can happen in any order):

 1. client sets command to run via proxy in sync mode
 2. docker creates exec context and runs proxy
 3a. client attaches to exec, puts terminals into raw mode, and waits
     for proxy to announce ready
 3b. proxy sets terminal to raw mode & disables echo, and then writes
     "PROXY READY\n" to stdout
 4. client sends terminal size to docker, and sets up resize forwarder
 5. client writes "PROXY RUN\n" to proxy stdin
 6a. proxy restores terminal and execs command
 6b. client connects stdin/stdout/stderr and waits for command to finish

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>