Switch from env whitelist to blacklist
It would be usefult to be able to pass environment variables through to the
scripts and commands running inside the container - so switch from a whitelist
to a blacklist.

The next step is to allow the project and user to manipulate the blacklist
(i.e. by adding or removing items).

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>