Support the old label names
If we don't read the old label names, then transition means breaking any
existing containers. So add support for reading the old labels, and mapping
them to the new ones before leaving the docker lib. We can't change the labels
on the container unfortunately, so we can't transition that way.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>