Try and get groups setup properly
We want the user running in the container to have the groups we have specified.
In order to acheive this, we need root privileges (or at least the appropriate
capability). In order to get this setup, we change the way we get wr-init into
the container, so that we can make it setuid. We also change the proxy to setup
the groups as we want, but also to drop root.

We also add an option to the crate config that can specify groups to be copied
from the host to the container (to facilitate mounting from the host to the
container).

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>