Install wharfrat and wr into the continaer
Currently this doesn't add any specific extra functionality for running inside
the container - but it's the first step.

Next we want to recognise when we have been requested to run in the current
container, and make that a no-op (i.e. "wr foo" would be the same as "foo" if
we were already in the appropriate container).

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>