Setup so symlinks to wharfrat will run commands
Now you can create a symlink to wharfrat as a shortcut for "wr <name>" - so you
have have a command on the host that will automatically run the same command
inside the container.

Option parsing is currently missing, but the intention is that you can set
WHARFRAT_OPTIONS to specify command line options to be parsed as if "wr" had
been called.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>