Fix error when running against missing image
If wr is run for an image that has been downloaded, then we will fail to get
the labels - which will cause immediate failure. However, we want to keep going
so that we can pull down the image when we try to create the container. So when
the error is a non-existant image, we simply return no labels, instead of an
error.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>