Recalculate defaults after pulling image
Since the crate defaults now can depend on information from the image, we need
to recalculate them after pulling the image. We also need to re-json the crate
in case those defaults have changed.

It may be better to change the way that create works, so that the entire
creation is retried if we pull the image - especially if more settings are
going to be based on information from the image.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>