Check container image matches configured one
Add a check that the image the existing container was built from is the correct
one. This means that either --clean or --force is necessary to run a container
where the image has been rebuilt.

Ideally the image should be a tagged image that doesn't change, but if you are
still developing the image it can be annoying to accidentally retest the same
container.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>