Add the ability to expose a command
The new wr-exec command provides the ability to expose a command from inside
the container to the host. It does this using a new config file in toml syntax,
the path to which is the first argument. Additional arguments are passed
through to the command run in the container (unless overridden by the config).
This allows creating files with '#!' lines that expose commands.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>