Add release script
Add a script to create a release tag. We update the default version information
as part of the process - so that building from the tag will display the version
information. We also merge the tag to the release branch. The release branch
should allow "go get" to download and build the tagged releases including the
correct version information.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>