Change the way local config works
Instead of having a single set of local setup instructions that apply to all
crates, have a list of setups - with optional project and crate regexs that can
be used to decide which applies.

Currently only the first entry that matches will be applied. It might actually
be better to apply them all.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>