Assign containers to users
When we create a container, include the username in the name. Also add the
owner's username as a label. The intention is not to prevent the user from
deliberately accesing containers for another user, but to maintain focus on
their own containers and prevent accidental access to containers of other
users.

Previously list would list containers from other users, which would then
potentially involve trying to access the project files, which may well be
inaccessible to the current user - much better to filter the container list to
only include the current user.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>