Catch failure to get "PROXY READY" response
If the command that we run fails before even starting the proxy, then we
would get stuck trying to set the terminal size - since we weren't
checking the reponse that we were getting to see if it actually came
from the proxy. So now we check to see if we got the expected reponse,
and exit with an error if we didn't.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>