Implement rebuild properly
Instead of deleting the venv and the trying to recreate it, we now only
rebuild the state that relates to the crate that has been recreated.
This now includes re-running the commands that we recorded were used to
install the exported binaries.

This should mean that if the crate needs to be rebuilt that we end up
with all of the export binaries still in a working state.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>