Display server version at startup
When the server starts inside the container, display the version
details.

Also, some tweaks to the reap setup to hopefully make it more reliable.
Since we have nothing else to do using signals seems unnecessary, so
just sit in a loop calling Wait4, which should block until a child
exits.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>