Specify entrypoint when creating container
When we are creating a new container for use with wharfrat we always
want to run wr-init as the main binary. Previously we were only
specifying Cmd when creating the container - which means that if the
image being used had an entrypoint set, that entrypoint would be used
and the container wouldn't be created as we expect. So now we specify
both the Entrypoint (though empty) and Cmd, which should mean that we
are in full control of what is running inside the container.

Note that we use an empty entrypoint, as we don't want the entrypoint
set to be used when we are trying to exec commands in the created
container.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>