Update create_group to support busybox
The setup_group method of Setup has been updated to handle using a
busybox environment where we only have a limited addgroup command, but
create_group had not. So update create_group with a similar approach so
that it can use a different command when running in a busybox based
container.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>