Start adjusting build script to work on mac
We need to make the build more complicated to accomodate mac. We need to
explictly tell Go that we want a mac executable since we are building in
a Linux container.

In addition, trying to turn CGO off doesn't really work with mac (and we
may need to use cgo anyway).

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>