Mac builds need access to a Linux build
When the native wharfrat is a static Linux executable then it can simply
copy itself to use inside the container. However, we can't run the Mac
version inside the container, so for the Mac we build a Linux version
and embed it into the Mac version. This makes the Mac version larger,
but makes life a bit simpler.

Maybe revisit this later?

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>