Don't start all crates when creating venv state
If a crate doesn't export any binaries, then there is no need to try and
start it when setting up a new venv (since it won't actually contribute
anything to the venv anyway).

This helps when the project file refers to crates that might not always
be working (e.g. where the image isn't available for all platforms).

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>