Delay output until logging is enabled
If we are running in sync mode then logging is disabled until the sync
is complete, so there is no point in trying to output anything before
the sync is complete. So move the startup output to after the sync is
complete - if sync is not enabled then the output happens at the same
time as it did before.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>