Add library for muxing traffic
I want to add the ability for an exec'd process to connect through an
arbitrary number of FDs to it's parent. Specifically I want to be able
to run delve dap using the exec mechanism from vscode, which means that
we need to connect FD 3 from vscode to dlv.

Since we only have a limited set of streams through docker, we will need
to use a mux to pass the traffic of the extra FDs that we want to
connect through a single connection each way.

We will also need some way to terminate the muxed connection on the
inside of the container to break the mux back out into a set of FDs.
This setup is going to have to involve the proxy staying running rather
than execing the command I suspect. Or perhaps we can get the init
process to terminate the mux somehow?

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>