Get the mux unit tests passing
Setup the write so that we can return an error from the write on the
other side. So now SplitCopy sends back the result of writing the data
so that the Writer's Write method can relay the actual result. This
combines with the already present close support to provide the ability
for two Conns to communicate not just data but also state information.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>