Try using socketpair to connect extra FDs
The problem with connecting arbitrary extra FDs other than
stdin/stdout/stderr is that we need a bidirectional connection since we
don't know which way they will be used. So we try using socketpair
instead of pipe to create a bidirectional connection. However, this
means that we need to do a bunch of low-level fiddling to get things
setup. I also haven't yet tested that it actually works as a
bidirectional pipe.