Fix issue with wr-init printing error
I couldn't figure out where the error was coming from for ages, but
eventually remembered that the flags package prints the error when the
default flags are used. So change the flags so that it doesn't do this,
and add explicit code to print the error when it's a help flag error.