Change internal command to use single dir close
In the same way that we had to change the external command to use halve
close commands instead of closing a bi-directional connection when one
way stopped, we need to do the same for the internal command. We now use
the halve close methods of the mux Connection, and the Shutdown function
to only close the connection in the direction that has stopped copying,
instead of closing down the entire connection nut only in one of the
parts.