Update Go, and docker module to 25.0.14
This isn't the latest version of the docker module, but it is the oldest
version that is compatible with the current version of docker API wise
(i.e. this the oldest version of docker that builds a working wharfrat).
Now that wharfrat is working again, then updating to the latest docker
becomes easier, as the dev environment is working again.

Updating docker and dependencies requires a newer version of Go, so also
update the newest Go current available.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>