Address some deprecation in docker module
Some of the types have been moved, and the old names are now deprecated.
So update to use the new names before the disappear.

Also, it would seem that a struct member has been deprecated, but it was
the wrong thing to have been using anyway.

Also fix some ignore error warnings from the linter, as it makes it
easier to see the deprecation warnings.

Signed-off-by: Julian Phillips <julian@quantumfyre.co.uk>